requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
umap-learn==0.5.3
scipy==1.10.1
